# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  events.criarEventoRequest:
    properties:
      date:
        type: string
      description:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
  integrations.criarEventoGoogleRequest:
    properties:
      date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  integrations.criarEventoOutlookRequest:
    properties:
      date:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
...
host: localhost:3001
info:
  contact: {}
  description: 'API do BeePlanner — autenticação via JWT (Bearer). Todas as rotas
    /api/* (exceto /auth) exigem o header Authorization: Bearer <token>.'
  title: BeePlanner API
  version: "1.0"
paths:
...
  /events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar eventos
      tags:
      - Events
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do evento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/events.criarEventoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar evento
      tags:
      - Events
...
  /integrations/google:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Desconectar Google Calendar
      tags:
      - Integrations
  /integrations/google/connect:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Iniciar conexão com Google Calendar
      tags:
      - Integrations
  /integrations/google/events:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - BearerAuth: []
      summary: Listar eventos do Google Calendar
      tags:
      - Integrations
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do evento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/integrations.criarEventoGoogleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar evento no Google Calendar
      tags:
      - Integrations
  /integrations/outlook:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Desconectar Outlook
      tags:
      - Integrations
  /integrations/outlook/connect:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Iniciar conexão com Outlook Calendar
      tags:
      - Integrations
  /integrations/outlook/events:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados do evento
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/integrations.criarEventoOutlookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Criar evento no Outlook Calendar
      tags:
      - Integrations
...
securityDefinitions:
  BearerAuth:
    description: 'Formato: Bearer <seu_jwt>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"